+++ mkdocs.yml
@@ -1,32 +1,38 @@
-site_name: Siddhi IO Hl7
-site_description: Siddhi IO Hl7 Extension
-site_author: WSO2
-site_url: https://wso2-extensions.github.io/siddhi-io-hl7/
-extra_css:
-  - stylesheets/extra.css
-extra_javascript:
-  - javascripts/extra.js
-repo_name: Siddhi-IO-hl7
-repo_url: https://github.com/wso2-extensions/siddhi-io-hl7
-copyright: Copyright &copy; 2011 - 2019 WSO2
+site_name: Siddhi IO HL7
+site_description: Siddhi IO HL7 Extension
+repo_name: Siddhi-IO-HL7
+repo_url: https://github.com/siddhi.io/siddhi-io-hl7
+edit_uri: https://github.com/siddhi.io/siddhi-io-hl7/blob/master
+copyright: Siddhi - Documentation
 theme:
   name: material
-  logo: images/siddhi-logo-w.svg
+  logo: images/siddhi-logo.svg
   favicon: images/favicon.ico
   palette:
     primary: teal
     accent: teal
-google_analytics:
-  - UA-103065-28
-  - auto
+extra_css:
+  - assets/stylesheets/extra.css
+extra_javascript:
+  - assets/javascripts/extra.js
 extra:
   social:
     - type: github
-      link: https://github.com/wso2/siddhi
+      link: https://github.com/siddhi-io/siddhi
+    - type: medium
+      link: https://medium.com/siddhi-io
+    - type: twitter
+      link: https://twitter.com/siddhi_io
     - type: linkedin
       link: https://www.linkedin.com/groups/13553064
+google_analytics:
+  - UA-103065-28
+  - auto
+markdown_extensions:
+  - markdown.extensions.admonition
 pages:
-- Welcome: index.md
+- Information: index.md
+
 - API Docs:
   - latest: api/latest.md
 - License: license.md
