+++ pom.xml
@@ -25,9 +25,9 @@
     <modules>
         <module>component</module>
     </modules>
-    <groupId>org.wso2.extension.siddhi.io.hl7</groupId>
+    <groupId>io.siddhi.extension.io.hl7</groupId>
     <artifactId>siddhi-io-hl7-parent</artifactId>
-    <version>2.0.2-SNAPSHOT</version>
+    <version>2.1.0-SNAPSHOT</version>
     <name>Siddhi I/O hl7</name>
 
     <scm>
@@ -119,19 +119,19 @@
                 </exclusions>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.text</groupId>
+                <groupId>io.siddhi.extension.map.text</groupId>
                 <artifactId>siddhi-map-text</artifactId>
                 <version>${siddhi.map.text.version}</version>
                 <scope>test</scope>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.xml</groupId>
+                <groupId>io.siddhi.extension.map.xml</groupId>
                 <artifactId>siddhi-map-xml</artifactId>
                 <version>${siddhi.map.xml.version}</version>
                 <scope>test</scope>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.json</groupId>
+                <groupId>io.siddhi.extension.map.json</groupId>
                 <artifactId>siddhi-map-json</artifactId>
                 <version>${siddhi.map.json.version}</version>
                 <scope>test</scope>
@@ -145,13 +145,13 @@
         </dependencies>
     </dependencyManagement>
     <properties>
-        <siddhi.version>5.0.0</siddhi.version>
+        <siddhi.version>5.1.2</siddhi.version>
         <siddhi.version.range>[5.0.0,6.0.0)</siddhi.version.range>
         <log4j.version>1.2.17.wso2v1</log4j.version>
         <carbon.transport.version>4.4.15</carbon.transport.version>
-        <siddhi.map.json.version>5.0.1</siddhi.map.json.version>
-        <siddhi.map.text.version>2.0.0</siddhi.map.text.version>
-        <siddhi.map.xml.version>5.0.0</siddhi.map.xml.version>
+        <siddhi.map.json.version>5.0.4</siddhi.map.json.version>
+        <siddhi.map.text.version>2.0.4</siddhi.map.text.version>
+        <siddhi.map.xml.version>5.0.3</siddhi.map.xml.version>
         <jacoco.plugin.version>0.7.9</jacoco.plugin.version>
         <testng.version>6.11</testng.version>
         <hapi.base.version>2.2</hapi.base.version>
